LayoutTests/media/progress-event-at-least-one-expected.txt

Test that at least one 'progress' event is fired before the 'load' event.

RUN(mediaElement = document.createElement('audio'))
RUN(mediaElement.src = 'content/test.wav')
RUN(mediaElement.load())

EVENT(loadstart)
EVENT(durationchange)
EVENT(loadedmetadata)
EVENT(loadeddata)
EVENT(canplaythrough)
EXPECTED (progressEventCount >= '1') OK
END OF TEST
